requirements.txt
streamlit
pandas
numpy
matplotlib
plotly
google-generativeai
openai
python-dotenv